+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Add get sessions command: `driver.sessions.get()`
 - Add get status: `driver.status.get()`
 - Add get capabilities: `driver.session.getCapabilities()`
+- Add push/pull files and folder: `driver.device.pushFile`, `driver.device.pullFile`, `driver.device.pullFolder`
 
 ## [0.0.6] - 2019-06-30
 - Add device lock related commands: `driver.device.*`